Summarize the reconstructed time series
Calculate the summary statistics of the reconstructed time series, then combine the results to descriptive table of the input dataset. The definition of 'n_claim' is different between 'ts_recon' and 'res_tbl_stat'. In 'ts_recon', the 'n_claim' represents the total number of claim for a particular drug, so one group may have multiple entries in one day. In 'res_tbl_stat', the 'n_claim' represents the number of unique patient claiming for drugs from a particular group throughout the years of observation. We will need to revisit this difference later and clarify the documentation.